Load data to be displayed from disk
Starting with imaginary test data.

I chose TypeScript to encode the product data because it provides type checking,
refactoring, inline comments, inline documentation with code assistance, and is
easy to slice and dice. It is hard to find the same combined functionality in
any of the usual data formats like YAML or JSON. XML might be the closest, but
that would require additional libraries and XML support on Node.js is rather
poor.