Introduce React Context
A context allows "global" state scoped to a specific NextJS page. Like Redux,
the context holds data and allows to mutate it safely with reducers. That makes
it easier to update the product comparison when the user applies filters and
requires fewer props to be passed around. Because the context can be provided
to arbitrary components, less boilerplate is needed to test them.

Using Redux would have required another library without providing more value.
Furthermore, Redux does not work well with NextJS.