Move filters to the front-end, make them filter
Making the filters applicable revealed that it would be beneficial to turn them
into classes containing data and behaviour (= methods), not only data. That
means that the filters have to be declared and created on the front-end,
because JSON (that Next.js uses to move data from the back-end to the
front-end) cannot carry functions.