Do not display back references in footnotes
Reusing footnotes requires multiple back references. Unfortunately, they become
useless quickly: How does the user figure out which one is the right back
reference to the previous position?

There are more questions: What happens if the user clicks on a back reference to
a product that is currently filtered out? Are multiple back references helpful
if they all point to the same row (reuse typically happens for a single feature
across multiple products of a vendor)?