Add production data
During build, the site either reads production data or test data based on the
environment variable DATA_SOURCE. Which data set is selected is controlled
through .env files.

During development and testing, test data is used because it provides a stable
foundation that also allows to test all aspects of the functionality which is
not necessarily the case with production data. To ensure that swapping data
sources works correctly and that the production build behaves correctly, a
number of acceptance tests is included.