Revert to using xargs
Another case of "I should have studied the man pages". I dropped xargs because I
did not want it to cause failures in case lsof did not find a matching process
(for example, because the server has crashed). But there is a command line
switch to instruct xargs not to run the utility if there is no match. And BSD
does the right thing anyway, even without the argument.