Replace mobile navigation
Making the main navigation fixed also required a new slide-out navigation on
small screens. The reason is that the old one relied on shifting the page
content down to display the navigation options below the main navigation bar.
Due to the navigation bar being fixed, this does no longer work.