Only terminate processes listening
lsof -i lists all connection states. As a consequence, the process list includes
processes that are connected to to a specific port, not only those that are
listening. Hence, the test script killed not only servers, but also clients,
including my browser if it were connected to the Next.js development server.