Fix broken opening and closing of filters
headlessui/react v2.1.3 broke transitions. Consequently, the popovers with the
filters would stop opening and closing after having been closed for the first
time. Downgrading to v2.1.2 fixes the problem.

I added two tests that verify that repeatedly opening and closing filters is
working fine. Further, I had to change the browser used by Cypress to Electron
because Cypress could somehow not detect that it is working correctly in
Firefox.