Replace Cypress in component tests
I want to get rid of Cypress because it breaks from time to time, is hard to
debug, and does not make it easy to write tests that are decoupled from the
implementation. The first step is to rewrite the existing component tests with
React Testing Library while keeping the old tests to ensure that I do not
introduce any bugs.